@@ -20,9 +20,9 @@ package org.esupportail.papercut.dao;
 import org.esupportail.papercut.domain.PayPapercutTransactionLog;
 import org.springframework.stereotype.Service;
 
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.persistence.Query;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
+import jakarta.persistence.Query;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Date;
