@@ -17,38 +17,37 @@
  */
 package org.esupportail.papercut.security;
 
-import java.io.IOException;
-
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.FilterConfig;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
-
+import jakarta.servlet.*;
+import jakarta.servlet.http.HttpServletRequest;
 import org.esupportail.papercut.web.WebUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
+import org.springframework.util.StringUtils;
+
+import java.io.IOException;
 
 @Component
 public class ContextFilter  implements Filter {
 
+	Logger log = LoggerFactory.getLogger(ContextFilter.class);
+
 	@Override
 	public void init(FilterConfig filterConfig) throws ServletException {
+		log.trace("ContextFilter initialized");
 	}
 
 	@Override
 	public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
 		HttpServletRequest request = (HttpServletRequest) servletRequest;
 		String context = WebUtils.getContext(request);
-		if (context != null && !context.isEmpty()) {
+		if (StringUtils.hasLength(context)) {
 			ContextHelper.setCurrentContext(context);
+			log.trace("ContextFilter set context: {} for url {}", context, request.getRequestURI());
 		} 
 		filterChain.doFilter(servletRequest, servletResponse);
+		log.trace("ContextFilter finished - clearing context");
 		ContextHelper.clear();
 	}
 
-	@Override
-	public void destroy() {
-	}
 }
