@@ -22,7 +22,7 @@ import org.esupportail.papercut.domain.PayPapercutTransactionLog;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 
 @Service
 public class AnonymizeOneService {
