@@ -17,21 +17,21 @@
  */
 package org.esupportail.papercut.web;
 
-import javax.annotation.Resource;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.annotation.Resource;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
 import org.esupportail.papercut.config.EsupPapercutConfig;
 import org.esupportail.papercut.config.EsupPapercutContext;
 import org.esupportail.papercut.security.ContextHelper;
 import org.esupportail.papercut.services.EsupPaperCutService;
+import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.ModelAndView;
-import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
 import org.springframework.web.servlet.view.RedirectView;
 import org.springframework.web.servlet.view.UrlBasedViewResolver;
 
-public class WebInterceptor extends HandlerInterceptorAdapter {
-	
+public class WebInterceptor implements HandlerInterceptor {
+
 	@Resource
 	EsupPapercutConfig config;
 	
@@ -45,34 +45,29 @@ public class WebInterceptor extends HandlerInterceptorAdapter {
 		String context = WebUtils.getContext(request);
 
 		if("webjars".equals(context) || "resources".equals(context)) {
-			super.postHandle(request, response, handler, modelAndView);
-		} else {
-
-			ContextHelper.setCurrentContext(context);
-
-			super.postHandle(request, response, handler, modelAndView);
-
-			if (modelAndView != null) {
-				boolean isViewObject = modelAndView.getView() == null;
-				boolean isRedirectView = !isViewObject && modelAndView.getView() instanceof RedirectView;
-				boolean viewNameStartsWithRedirect = isViewObject && modelAndView.getViewName().startsWith(UrlBasedViewResolver.REDIRECT_URL_PREFIX);
-
-				if (!isRedirectView && !viewNameStartsWithRedirect && context != null) {
-					EsupPapercutContext configContext = config.getContext(context);
-					if (configContext != null) {
-						modelAndView.addObject("title", configContext.getTitle());
-						modelAndView.addObject("htmlFooter", configContext.getHtmlFooter());
-						modelAndView.addObject("pContext", context);
-					}
-					modelAndView.addObject("isAdmin", WebUtils.isAdmin());
-					modelAndView.addObject("isManager", WebUtils.isManager());
-					modelAndView.addObject("availableContexts", WebUtils.availableContexts());
-					modelAndView.addObject("payAvailable", !esupPaperCutService.getPayModes(configContext).isEmpty());
-				}
-			}
+			return;
 		}
 
+		ContextHelper.setCurrentContext(context);
+
+		if (modelAndView != null) {
+			boolean isViewObject = modelAndView.getView() == null;
+			boolean isRedirectView = !isViewObject && modelAndView.getView() instanceof RedirectView;
+			boolean viewNameStartsWithRedirect = isViewObject && modelAndView.getViewName().startsWith(UrlBasedViewResolver.REDIRECT_URL_PREFIX);
+
+			if (!isRedirectView && !viewNameStartsWithRedirect && context != null) {
+				EsupPapercutContext configContext = config.getContext(context);
+				if (configContext != null) {
+					modelAndView.addObject("title", configContext.getTitle());
+					modelAndView.addObject("htmlFooter", configContext.getHtmlFooter());
+					modelAndView.addObject("pContext", context);
+				}
+				modelAndView.addObject("isAdmin", WebUtils.isAdmin());
+				modelAndView.addObject("isManager", WebUtils.isManager());
+				modelAndView.addObject("availableContexts", WebUtils.availableContexts());
+				modelAndView.addObject("payAvailable", !esupPaperCutService.getPayModes(configContext).isEmpty());
+			}
+		}
 	}
 
 }
-
